+++ packages/server/__tests__/lib/test-global-setup.js
@@ -1,0 +1,10 @@
+const db = require('../../src/db');
+const fixtures = require('../db/seeds/fixtures');
+
+before(async () => {
+    await fixtures.seed(db.knex);
+});
+
+after(async () => {
+    await db.knex.destroy();
+});
